+++ addOns/openapi/src/main/java/org/zaproxy/zap/extension/openapi/spider/OpenApiSpider.java
@@ -19,26 +19,75 @@
  */
 package org.zaproxy.zap.extension.openapi.spider;
 
+import java.util.Locale;
 import java.util.function.Supplier;
+import org.apache.commons.lang.StringUtils;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.parosproxy.paros.network.HttpHeader;
+import org.parosproxy.paros.network.HttpMessage;
+import org.parosproxy.paros.network.HttpSender;
 import org.zaproxy.addon.spider.parser.ParseContext;
 import org.zaproxy.addon.spider.parser.SpiderParser;
+import org.zaproxy.zap.extension.openapi.HistoryPersister;
+import org.zaproxy.zap.extension.openapi.converter.Converter;
+import org.zaproxy.zap.extension.openapi.converter.swagger.SwaggerConverter;
+import org.zaproxy.zap.extension.openapi.network.Requestor;
 import org.zaproxy.zap.model.ValueGenerator;
 
 public class OpenApiSpider extends SpiderParser {
 
-    private OpenApiSpiderFunctionality func;
+    private static final Logger LOGGER = LogManager.getLogger(OpenApiSpider.class);
+    private Requestor requestor;
+    private Supplier<ValueGenerator> valGenSupplier;
 
     public OpenApiSpider(Supplier<ValueGenerator> valueGeneratorSupplier) {
-        func = new OpenApiSpiderFunctionality(valueGeneratorSupplier);
+        valGenSupplier = valueGeneratorSupplier;
+        requestor = new Requestor(HttpSender.SPIDER_INITIATOR);
+        requestor.addListener(new HistoryPersister());
     }
 
     @Override
     public boolean parseResource(ParseContext ctx) {
-        return func.parseResource(ctx.getHttpMessage());
+        HttpMessage message = ctx.getHttpMessage();
+        try {
+            Converter converter =
+                    new SwaggerConverter(
+                            null,
+                            message.getRequestHeader().getURI().toString(),
+                            message.getResponseBody().toString(),
+                            valGenSupplier.get());
+            requestor.run(converter.getRequestModels());
+        } catch (Exception e) {
+            LOGGER.debug(e.getMessage(), e);
+            return false;
+        }
+
+        return true;
     }
 
     @Override
     public boolean canParseResource(ParseContext ctx, boolean wasAlreadyConsumed) {
-        return func.canParseResource(ctx.getHttpMessage());
+        HttpMessage message = ctx.getHttpMessage();
+        try {
+            String contentType =
+                    message.getResponseHeader()
+                            .getHeader(HttpHeader.CONTENT_TYPE)
+                            .toLowerCase(Locale.ROOT);
+            String responseBodyStart =
+                    StringUtils.left(message.getResponseBody().toString(), 250)
+                            .toLowerCase(Locale.ROOT);
+            if (contentType.startsWith("application/vnd.oai.openapi")
+                    || (contentType.contains("json")
+                            || contentType.contains("yaml")
+                                    && (responseBodyStart.contains("swagger")
+                                            || responseBodyStart.contains("openapi")))) {
+                return true;
+            }
+        } catch (Exception e) {
+            return false;
+        }
+        LOGGER.debug("Can't parse {}", message.getRequestHeader().getURI());
+        return false;
     }
 }